Use bulk SPI writes and support 128x64 displays
SPI.transact() is destructive, but we get around that by using a separate
write buffer. This is needed to support 128x64 displays anyways, since
they require sending the 4-bit pixel values twice.

This surely makes it stop working for 256x64 displays, but I don't have one
of those to test with.